% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/murlib.tidyverse-package.R
\docType{package}
\name{murlib.tidyverse-package}
\alias{murlib.tidyverse}
\alias{murlib.tidyverse-package}
\title{murlib.tidyverse: Tidyverse Enhancements}
\description{
Various functions and objects that build upon {tidyverse} constructs.
}
\author{
\strong{Maintainer}: Murat Tasan \email{mmuurr@gmail.com} (\href{https://orcid.org/0000-0002-8155-2565}{ORCID})

}
\keyword{internal}
